PDF Summarizer/requirements.txt
flask==2.3.2
openai==0.27.0
python-dotenv==1.0.0
PyPDF2==3.0.0
werkzeug==2.3.2
